+++ bzwgen/inc/Product.h
@@ -20,10 +20,10 @@
 
 class Product {
   OperationVector* ops;
-  float rarity;
+  double rarity;
   Expression* cond;
 public:
-  Product(OperationVector* _ops, float _rarity, Expression* _cond = NULL) : ops(_ops), rarity(_rarity), cond(_cond) {};
+  Product(OperationVector* _ops, double _rarity, Expression* _cond = NULL) : ops(_ops), rarity(_rarity), cond(_cond) {};
   int runMesh(Mesh* mesh, int face) {
     for (size_t i = 0; i < ops->size(); i++) {
       face = ops->at(i)->runMesh(mesh,face);
@@ -32,9 +32,9 @@ public:
     return face;
   }
   bool conditionsMet(Mesh* mesh, int face) { 
-    return cond == NULL ? true : (cond->calculate(mesh,face) >= 0.0f);
+    return cond == NULL ? true : (cond->calculate(mesh,face) >= 0.0);
   }
-  float getRarity() { return rarity; }
+  double getRarity() { return rarity; }
   ~Product() {
     OperationVectIter itr; 
     for (itr = ops->begin(); itr!= ops->end(); ++itr) delete (*itr);
